#68 スコア詳細でソートすると全体にレンダリングがかかる問題の修正
